x-dm-bot/requirements.txt
# Core dependencies
selenium==4.18.1
undetected-chromedriver==3.5.5
python-dotenv==1.0.1
schedule==1.2.0
cryptography==42.0.2

# Web GUI dependencies
flask==3.0.0
flask-cors==4.0.0

# Chrome driver management
webdriver-manager==4.0.1

# Data handling
pandas==2.2.0
openpyxl==3.1.2

# Logging and monitoring
colorlog==6.8.2

# Rate limiting and retry
tenacity==8.2.3
ratelimit==2.2.1

# HTTP requests
requests==2.31.0
urllib3==2.2.0

# Optional: For advanced features
# selenium-stealth==1.0.6
# selenium-wire==5.1.0  # For proxy support with authentication

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-mock==3.12.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Type hints
types-requests==2.31.0.20240125
types-python-dateutil==2.8.19.20240106
# 自動開発システム用パッケージ
dash
plotly
psutil
radon
alembic
sqlalchemy
pytest
pytest-cov
black
flake8
websocket-client
celery
redis
